+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @thaparoyal/calendar-core
 
+## 2.0.0
+
+### Major Changes
+
+- Updated some packages
+
 ## 1.0.0
 
 ### Major Changes